♻️ (aa-policy): Extract policy document semantics into a leaf crate
AAASM-5349 requires audit, `status` and `verify` to report which of the
four policy states a governed launch resolved. `status` and `verify` are
rendered from what the developer-integration service in `aa-runtime`
computed — ADR 0030 forbidden design 10 prohibits the client deriving a
protection state for itself — so the service must be able to resolve a
policy. It could not: `aa-runtime` does not depend on `aa-gateway`, and
must not, because the runtime is the enforcement point and the gateway
is the brain above it.

Rather than invert that layering, the semantics move below both. Parsing,
validation, the document model, the expression evaluator and version
history have no gateway dependencies at all, so they become `aa-policy`,
a leaf beside `aa-core` and `aa-security`. `aa-gateway::policy` re-exports
it wholesale, so every existing `aa_gateway::policy::…` path keeps
resolving and no caller changes.

The split falls where the state does. `PolicyContext` — the contract for
topology-aware condition variables — moves with the evaluator that calls
it, since it names nothing but `aa_core` types. Its production
implementation stays in `aa-gateway`, which owns the `AgentRegistry` and
`BudgetTracker` it reads. That one boundary is the whole reason the leaf
crate is possible.

Pure move: no behaviour, no signature and no test assertion changes.
Three items widen visibility only because a crate boundary now sits where
a module boundary did — `canonical`, `expr` and `references_live_context`
were `pub(crate)` for callers that are still the same callers.
