+++ aa-gateway/src/policy/mod.rs
@@ -1,43 +1,21 @@
-//! Policy YAML parser and validator for aa-gateway.
+//! Policy semantics as the gateway sees them: the shared document model from
+//! [`aa_policy`], plus the evaluation context only this crate can supply.
 //!
-//! Entry point: [`validator::PolicyValidator::from_yaml`].
+//! # Why the split (AAASM-5349)
 //!
-//! # Single canonical AST (AAASM-3607)
+//! Parsing and validating a policy document is not a gateway concern — the
+//! developer-integration service in `aa-runtime` and `aasm run` must reach the
+//! same answer about which policy is in effect, and neither may depend on the
+//! gateway. Those semantics therefore live in the leaf crate [`aa_policy`] and
+//! are re-exported here unchanged, so `aa_gateway::policy::…` keeps meaning
+//! exactly what it did.
 //!
-//! The cross-layer-shared dimensions of a policy (capabilities, network
-//! egress, tool rules) are defined once in [`aa_security::policy`]. The gateway
-//! keeps its richer in-crate [`PolicyDocument`] for L7-only evaluation (CEL,
-//! history, budget) but projects onto the canonical AST via
-//! [`PolicyDocument::to_canonical`]. The eBPF kernel rules are lowered
-//! (`aa_security::policy::lower_to_ebpf`) from that same canonical projection,
-//! so the L7 engine and the kernel layer provably share one definition — there
-//! is no second, divergent copy of the shared schema.
+//! What stays is [`context`]: evaluating a policy needs a budget tracker and
+//! the agent registry, which are gateway state. That dependency is the reason
+//! the split falls where it does.
 
-pub(crate) mod canonical;
-pub(crate) mod context;
-pub mod digest;
-pub mod document;
-pub mod error;
-pub(crate) mod expr;
-pub mod history;
-pub mod network;
-pub mod raw;
-pub mod rbac;
-pub mod scope;
-pub mod validator;
+pub use aa_policy::*;
 
-pub use context::{ContextError, PolicyContext};
-pub use document::{ActiveHours, BudgetPolicy, DataPolicy, NetworkPolicy, PolicyDocument, SchedulePolicy, ToolPolicy};
-pub use error::{PolicyParseError, ValidationError, ValidationWarning};
-pub use expr::{evaluate_clause, ClauseKind, ResolutionFailure};
-pub use network::{check_network_egress, EgressDecision};
-pub use rbac::{required_role_for, CallerRole, MutationKind, PolicyScopeKind};
-pub use scope::{OrgId, PolicyScope, TeamId};
-pub use validator::{PolicyValidator, PolicyValidatorOutput};
+pub mod context;
 
-// Re-export the canonical, cross-layer policy AST so consumers of
-// `aa_gateway::policy` reach the single source of truth in `aa-security`.
-pub use aa_security::policy::{
-    lower_to_ebpf, Capability as CanonicalCapability, CapabilitySet as CanonicalCapabilitySet, EbpfRuleSet, PathRule,
-    PathVerdict, PolicyDocument as CanonicalPolicyDocument,
-};
+pub use context::ProductionPolicyContext;
