+++ aa-policy/src/history/snapshot.rs
